Complete functional tab completion interface
