--- src/main/java/de/seine_eloquenz/lbcfs/command/TabOption.java
@@ -1,6 +1,0 @@
-package de.seine_eloquenz.lbcfs.command;
-
-@FunctionalInterface
-public interface TabOption {
-    String[] get();
-}
